+++ src/components/Modals/CreateCardModal.tsx
@@ -37,7 +37,7 @@ import dayjs from 'dayjs';
 import Image from 'next/image';
 
 import ImageCropperModal from '@components/Modals/ImageCropperModal';
-import { DashBoardContext } from '@core/contexts/DashBoardContext';
+import { DashBoardContext } from '@core/contexts/DashboardContext';
 import { CreateCardRequestDto } from '@core/dtos/CardsDto';
 import addPurple from '@icons/add_purple.png';
 import calendar from '@icons/calendar.png';
